Deleted Matrix and Parameter Classes. Changed CalcVector to use array internally. This means adding new parameters at runtime is not possible anymore. Also the names of parameters are not stored inside the student object anymore. Please test all functions that use these objects and make changes where necessary.

+++ src/edu/dhbw/sos/course/student/IPlace.java
@@ -35,7 +35,7 @@ public interface IPlace {
 	void printAcutalState();
 	
 	
-	void donInput(int index, int value);
+	void donInput(int index, float value);
 	
 	
 	public LinkedHashMap<Integer, CalcVector> getHistoryStates();
